finished ui tests for product screen

+++ app/src/main/java/com/gmail/bodziowaty6978/fitnessappv2/feature_diary/domain/model/DiaryEntry.kt
@@ -13,6 +13,6 @@ data class DiaryEntry(
     val date: String = Date(System.currentTimeMillis()).toString(),
     val time:Long = System.currentTimeMillis(),
     val brand: String? = "",
-    val weight: Double = 0.0,
+    val weight: Int = 0,
     val unit: String = "g",
     val nutritionValues: NutritionValues = NutritionValues()):Parcelable
